Add examples and fix bug in common.h

+++ include/common.h
@@ -3,6 +3,6 @@
 
 int cmpint(const void *a, const void *b);
 
-double cmpdouble(const void *a, const void *b);
+int cmpdouble(const void *a, const void *b);
 
 #endif
